requirements.txt
pytesseract
pdf2image
PyMuPDF
transformers
accelerate
pillow
requests
python-dotenv
google-genai
fastapi
torch
scipy
uvicorn
python-multipart
opencv-python
